update: - arranged external modules together, arranged the internal modules together, added a module for responses, added my twitter handle.
