+++ modules/usernames.js
@@ -9,4 +9,5 @@ module.exports = [
     '@theLiveware',
     '@efemoney_',
     '@udezekene',
+    '@ajafik',
 ];
